pyproject.toml
[tool.poetry]
name = "deeplda-project"
version = "0.1.0"
description = "DeepLDA implementation using ResNet and PyTorch on CIFAR-10"
authors = ["Your Name <you@example.com>"] # Optional: Change this
readme = "README.md" # Assuming you might have a README later
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"  # Requires Python 3.8 or higher
torch = "*"
torchvision = "*"
numpy = "*"
Pillow = "*"
tqdm = "*"
scikit-learn = "*"
wandb = "^0.19.9"
pyramid = "^2.0.2"
